added options

+++ main.py
@@ -80,21 +80,26 @@ def read_and_close_positions():
     # json_response = response.json()
     order_id_data = response_positions.json()
     print("\t Read orders data:", order_id_data)
-    order_ids = [order['id'] for order in order_id_data['orders']['order'] if order['status'] =='pending']
-    print("\torder ids=",order_ids)
-    # Cancel orders
-    for order_id in order_ids:
-        print("\tCanceling Order ID:",order_id)
-        response_delete = requests.delete(f'https://sandbox.tradier.com/v1/accounts/VA88823939/orders/{order_id}',
-                data={},
-                headers={'Authorization': 'Bearer pOPACO7fKI7Alz4hHIQB66jFDACP', 'Accept': 'application/json'}
-                                 )
-        print("\t Order delete status :", response_delete.json())
 
+    if order_id_data['orders'] == 'null':
+        print("\t No Positions found")
+    else:
+        order_ids = [order['id'] for order in order_id_data['orders']['order'] if order['status'] =='pending']
+        print("\torder ids=",order_ids)
+        # Cancel orders
+        for order_id in order_ids:
+            print("\tCanceling Order ID:",order_id)
+            response_delete = requests.delete(f'https://sandbox.tradier.com/v1/accounts/VA88823939/orders/{order_id}',
+                    data={},
+                    headers={'Authorization': 'Bearer pOPACO7fKI7Alz4hHIQB66jFDACP', 'Accept': 'application/json'}
+                                     )
+            print("\t Order delete status :", response_delete.json())
+
+    # read positions
     response_positions = requests.get('https://sandbox.tradier.com/v1/accounts/VA88823939/positions',
-                             params={},
-                             headers={'Authorization': 'Bearer pOPACO7fKI7Alz4hHIQB66jFDACP',
-                                      'Accept': 'application/json'}
+                    params={},
+                    headers={'Authorization': 'Bearer pOPACO7fKI7Alz4hHIQB66jFDACP',
+                                          'Accept': 'application/json'}
             )
     #json_response = response.json()
     print("\t Read data:", response_positions.json())
@@ -102,50 +107,32 @@ def read_and_close_positions():
     # Retrieve all existing positions
     order_id_data = response_positions.json()
     if response_positions.status_code == 200:
-        positions_data = response_positions.json().get('positions', [])
-        positions_data = [position['symbol'] for position in order_id_data['positions']['position'] if position['quantity'] > 0]
-        print("\t position data:", positions_data)
-        for position in positions_data:
-            if positions_data == 'null':
+        # positions_data = response_positions.json().get('positions', [])
+        # print(order_id_data)
+            if order_id_data['positions'] == 'null':
                 print("\t No Positions found")
             else:
-                symbol = position
-                quantity = 1
-                side = "sell_to_close" if quantity > 0 else "buy_to_close"  # Close position by selling if long, buying if short
-
-                # ### check any existing order for that positions
-                # response_positions = requests.get('https://sandbox.tradier.com/v1/accounts/VA88823939/orders',
-                #              params={},
-                #              headers={'Authorization': 'Bearer pOPACO7fKI7Alz4hHIQB66jFDACP',
-                #                       'Accept': 'application/json'}
-                # )
-                # # json_response = response.json()
-                # order_id_data = response_positions.json()
-                # print("\t Read orders data:", order_id_data)
-                # print("\t Read orders data2:", response_positions)
-                # # extract order ID for this symbol
-                # found_id = None
-                # for position in order_id_data.get('position', []):
-                #     if position['option_symbol'] == symbol:
-                #         found_id = position['id']
-                #         break
-                # print("\t Found Order ID for Symbol",symbol,found_id)
-                 # Place order to close position
-                print("\t Send order to close at Market price : ", symbol,quantity)
-                response = requests.put('https://sandbox.tradier.com/v1/accounts/VA88823939/orders/found_id',
-                                 data={'class': 'option',
-                                       'symbol': 'SPX',
-                                       'option_symbol': symbol,
-                                       'side': side,
-                                       'quantity': quantity,
-                                       'type': 'market',
-                                       'duration': 'day',
-                                       'tag': 'my-tag-example-1'},
-                                 headers={'Authorization': 'Bearer pOPACO7fKI7Alz4hHIQB66jFDACP',
-                                          'Accept': 'application/json'}
-                                 )
+                print("\t Read dataxxxxx:", order_id_data)
+                for position in order_id_data['positions']['position']:
+                    symbol = position['symbol']
+                    quantity = position['quantity']
+                    side = "sell_to_close" if quantity > 0 else "buy_to_close"  # Close position by selling if long, buying if short
+                    print("\t Send order to close at Market price : ", symbol,quantity)
+                    response = requests.post('https://sandbox.tradier.com/v1/accounts/VA88823939/orders',
+                                     data={'class': 'option',
+                                           'symbol': 'SPX',
+                                           'option_symbol': symbol,
+                                           'side': side,
+                                           'quantity': quantity,
+                                           'type': 'market',
+                                           'duration': 'day',
+                                           'tag': 'my-tag-example-1'},
+                                     headers={'Authorization': 'Bearer pOPACO7fKI7Alz4hHIQB66jFDACP',
+                                              'Accept': 'application/json'}
+                                     )
 
                 json_response = response.json()
+                #print("CLose positions",response.json())
 
 
                 if response.status_code == 201:
